Replaced size_t integers by uint64_t's in eval
(required to avoid overflow when processing large datasets)

+++ DatasetUtils.h
@@ -42,7 +42,7 @@ public:
 	const std::string m_sName;
 	const std::string m_sDBName;
 	std::vector<SequenceInfo*> m_vpSequences;
-	size_t nTP, nTN, nFP, nFN;
+	uint64_t nTP, nTN, nFP, nFN;
 };
 
 class SequenceInfo {
@@ -56,7 +56,7 @@ public:
 	cv::Mat GetSequenceROI() const;
 	const std::string m_sName;
 	const std::string m_sDBName;
-	size_t nTP, nTN, nFP, nFN;
+	uint64_t nTP, nTN, nFP, nFN;
 private:
 	std::vector<std::string> m_vsInputFramePaths;
 	std::vector<std::string> m_vsGTFramePaths;
@@ -89,7 +89,7 @@ static inline void WriteResult(	const std::string& sResultsPath,
 	cv::imwrite(sResultFilePath.str(), res, vnComprParams);
 }
 
-static inline void WriteMetrics(const std::string sResultsFileName, size_t nTP, size_t nTN, size_t nFP, size_t nFN) {
+static inline void WriteMetrics(const std::string sResultsFileName, uint64_t nTP, uint64_t nTN, uint64_t nFP, uint64_t nFN) {
 	std::ofstream oMetricsOutput(sResultsFileName);
 	oMetricsOutput << nTP << " " << nFP << " " << nFN << " " << nTN << std::endl; // order similar to the files saved by the CDNet analysis script
 	oMetricsOutput.close();
@@ -202,7 +202,7 @@ static inline void GetSubDirsFromDir(const std::string& sDirPath, std::vector<st
 #endif
 }
 
-static inline void CalcMetricsFromResult(const cv::Mat& oInputFrame, const cv::Mat& oGTFrame, const cv::Mat& oROI, size_t& nTP, size_t& nTN, size_t& nFP, size_t& nFN) {
+static inline void CalcMetricsFromResult(const cv::Mat& oInputFrame, const cv::Mat& oGTFrame, const cv::Mat& oROI, uint64_t& nTP, uint64_t& nTN, uint64_t& nFP, uint64_t& nFN) {
 	CV_DbgAssert(oInputFrame.type()==CV_8UC1 && oGTFrame.type()==CV_8UC1 && oROI.type()==CV_8UC1);
 	CV_DbgAssert(oInputFrame.size()==oGTFrame.size() && oInputFrame.size()==oROI.size());
 	const int step_row = oInputFrame.step.p[0];
